fix(ci): address PR #486 review feedback
- cancel: validate that --workflow belongs to <run-id> before issuing
  the cancel, so `depot ci cancel wrong_run --workflow=wf_real` no
  longer silently cancels wf_real. (cursor[bot], ischolten)
- dispatch: use https://depot.dev/orgs/{org}/workflows/{run_id} for
  the post-trigger link, matching `depot ci run`. (ischolten)
- proto(DispatchWorkflowRequest.workflow): correct the field comment
  -- backend wants the basename ("deploy.yml"), not the full repo
  path. The CLI's --workflow help text already said this; the proto
  is now consistent. (ischolten)
- resolve: format the available-workflows list with strings.Join so
  the error reads "available: wf_a, wf_b" instead of Go's "[wf_a wf_b]".
  Apply the same treatment in the new validateWorkflowInRun. (ischolten)
- output: extract a tiny writeJSON helper and use it from all four
  new verbs (six call sites collapsed). Sweep of the older verbs
  (run, secrets, vars, ssh) is intentionally out of scope for this
  PR. (ischolten)
- tests: move TestCICommandRegistration from rerun_test.go (where it
  was misplaced) into a dedicated ci_test.go, and expand it to guard
  all 11 expected subcommands instead of just the four new ones.
  (ischolten)

Refs: depot/cli#486
Made-with: Cursor
